修改告警的问题  Signed-off-by: shuxinyiA <zhangcui11@huawei.com>

+++ device/plugins/hidump_plugin/src/hidump_plugin.cpp
@@ -165,7 +165,7 @@ bool HidumpPlugin::ParseHidumpInfo(HidumpInfo& dataProto, char *buf, int len)
             }
             eve->set_fps(atoi(databuf));
             (void)memset_s(databuf, BUF_MAX_LEN, 0, BUF_MAX_LEN);
-            size_t bufLength = strlen(buf) - length - 1;
+            int bufLength = strlen(buf) - length - 1;
             if (bufLength > 0 && BUF_MAX_LEN > bufLength) {
                 if (memcpy_s(databuf, BUF_MAX_LEN, buf + length + 1, bufLength) != EOK) {
                     HILOG_ERROR(LOG_CORE, "copy %d byte to memory region [%p, %p) FAILED!",
